separate API client and Langchain wrapper.
